Type mistake corrected

+++ main.cpp
@@ -4,8 +4,8 @@ int main() {
 	int num_of_val = 0;
 	std::cin >> num_of_val;
 
-	int prev_val = 0;
-	int cur_val = 1;
+	unsigned long long int prev_val = 0;
+	unsigned long long int cur_val = 1;
 
 	if(num_of_val == 0) {
 		std::cout << 0 << std::endl;
@@ -13,7 +13,7 @@ int main() {
 	}
 
 	for(int i = 1; i < num_of_val; i++) {
-		int tmp = cur_val;
+		unsigned long long int tmp = cur_val;
 		cur_val += prev_val;
 		prev_val = tmp;
 	}
